--- a/src/txdb.h
+++ b/src/txdb.h
@@ -99,8 +99,8 @@ class CTransactionDB : public CTransactionDBView {
    public:
     bool SetTxCache(const uint256 &blockHash, const vector<uint256> &hashTxSet);
     bool GetTxCache(const uint256 &hashblock, vector<uint256> &hashTx);
-    bool LoadTransaction(map<uint256, vector<uint256> > &mapTxHashByBlockHash);
-    bool BatchWrite(const map<uint256, vector<uint256> > &mapTxHashByBlockHash);
+    bool LoadTransaction(map<uint256, set<uint256> > &mapTxHashByBlockHash);
+    bool BatchWrite(const map<uint256, set<uint256> > &mapTxHashByBlockHash);
     int64_t GetDbCount() { return db.GetDbCount(); }
 };